Modularize shared facet code
